Rewrite/comment tests of SIAE invitations
- tests_base contains only tests about SIAE (not the abstract)
- many tests don't properly test errors and because the guest isn't a SIAE member,
  they think to test another error and are happy with that.
- several tests don't test what they think to test
- bug in TOTAL_FORM so the real bug wasn't found
- don't detect duplicates in invitations
- don't test the extended date
- no full coverage of _extended_expiration_date_or_error
- many tests step on each other's toes

Add FIXME in waiting for real fixes because the diff is already large.
The diff is hard to read because it contains too many changes.

Files will be renamed in next commit because test_base was/is about SIAE.
